Move unit tests for DeltaSet
